Moving iree/base/file_io.h to iree/base/internal/. There's some goo here for internal support that I forget if is still used.
