+++ iree/base/internal/file_io_posix.cc
@@ -25,8 +25,8 @@
 #include <cstdlib>
 
 #include "absl/strings/str_cat.h"
-#include "iree/base/file_io.h"
 #include "iree/base/file_path.h"
+#include "iree/base/internal/file_io.h"
 #include "iree/base/status.h"
 #include "iree/base/tracing.h"
