+++ iree/base/internal/file_io_test.cc
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "iree/base/file_io.h"
+#include "iree/base/internal/file_io.h"
 
 #include "absl/strings/str_cat.h"
 #include "absl/strings/string_view.h"
