+++ iree/base/testing/BUILD
@@ -44,8 +44,8 @@ cc_test(
         ":dynamic_library_test_library",
         "//iree/base:core_headers",
         "//iree/base:dynamic_library",
-        "//iree/base:file_io",
         "//iree/base:status",
+        "//iree/base/internal:file_io",
         "//iree/testing:gtest",
         "//iree/testing:gtest_main",
     ],
