+++ iree/base/testing/CMakeLists.txt
@@ -53,7 +53,7 @@ iree_cc_test(
     ::dynamic_library_test_library
     iree::base::core_headers
     iree::base::dynamic_library
-    iree::base::file_io
+    iree::base::internal::file_io
     iree::base::status
     iree::testing::gtest
     iree::testing::gtest_main
