+++ iree/tools/iree-benchmark-module-main.cc
@@ -17,8 +17,8 @@
 #include "absl/flags/usage.h"
 #include "absl/strings/string_view.h"
 #include "benchmark/benchmark.h"
-#include "iree/base/file_io.h"
 #include "iree/base/flags.h"
+#include "iree/base/internal/file_io.h"
 #include "iree/base/status.h"
 #include "iree/base/tracing.h"
 #include "iree/hal/drivers/init.h"
