+++ iree/tools/iree-check-module-main.cc
@@ -18,8 +18,8 @@
 #include "absl/strings/match.h"
 #include "absl/strings/string_view.h"
 #include "iree/base/api.h"
-#include "iree/base/file_io.h"
 #include "iree/base/flags.h"
+#include "iree/base/internal/file_io.h"
 #include "iree/base/status.h"
 #include "iree/base/target_platform.h"
 #include "iree/base/tracing.h"
